let the user pass time instead of the time_type
Another update! I don't see a big usecase in specifying the time type,
but what really happens is that I forget to start the
timer. So you try to stop your timer and notice that you forgot to start
it, so it's very easy to remember because it's probably on the same
day (I assume :grin:). So you can start it at the time you started
working and then end again.